% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_refs.R
\name{anova_textbook}
\alias{anova_textbook}
\title{The MDK textbook citation}
\usage{
anova_textbook()
}
\value{
The full reference, as a character string.
}
\description{
The MDK textbook citation
}
\examples{
anova_textbook()
}
\keyword{documentation}
