% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_information.R
\name{cor_format}
\alias{cor_format}
\title{Generate a Nice-Looking Correlation Matrix with Custom Bolded Values}
\usage{
cor_format(cor_mat, bold_val = 0.3)
}
\arguments{
\item{cor_mat}{A numeric correlation matrix (data frame or matrix) with numeric values representing the correlation coefficients.}

\item{bold_val}{A numeric value that sets the threshold for bolding correlations. Correlations greater than this value (or less than the negative value) will be bolded.
default = .30}
}
\value{
A styled HTML table representing the correlation matrix with bolded values based on the specified threshold.
}
\description{
This function takes a correlation matrix and formats it into a nicely styled HTML table,
with the ability to bold correlation coefficients above and below a specified threshold (\code{bold_val}).
It supports bolding values greater than the threshold, less than the negative threshold, or equal to 1 (diagonal).
}
\details{
The function generates a correlation matrix where the correlation coefficients that exceed the specified \code{bold_val}
(in absolute value) are bolded. This can be useful for highlighting strong correlations in a matrix. The matrix is
displayed using \code{kableExtra} to produce a polished HTML table. Diagonal elements (which are always 1) are bolded by default,
but this can be adjusted by modifying the \code{ifelse} condition inside the \code{mutate()} function.
}
\examples{
# Example usage of cor_mat_bold_fun
cor_matrix <- cor(mtcars)
cor_format(cor_matrix, bold_val = 0.3)
}
