% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_assumptions.R
\name{.engine_oneway_bs}
\alias{.engine_oneway_bs}
\title{One-way between-subjects engine.}
\usage{
.engine_oneway_bs(model, center = "mean", alpha = 0.05)
}
\description{
One-way between-subjects engine.
}
\keyword{internal}
