% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_assumptions.R
\name{.test_homogeneity}
\alias{.test_homogeneity}
\title{Homogeneity of variance, as a finding.}
\usage{
.test_homogeneity(y, g, center = "mean", alpha = 0.05)
}
\description{
Homogeneity of variance, as a finding.
}
\keyword{internal}
