% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_assumptions.R
\name{.test_normality}
\alias{.test_normality}
\title{Normality of residuals, as a finding.}
\usage{
.test_normality(model, alpha = 0.05)
}
\description{
Normality of residuals, as a finding.
}
\keyword{internal}
