% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_histogram.R
\name{dynamic_histogram}
\alias{dynamic_histogram}
\title{Generate Histogram for a Given Variable}
\usage{
dynamic_histogram(df, variable)
}
\arguments{
\item{df}{A data frame containing the variable to be plotted.}

\item{variable}{The variable in the data frame for which the histogram is generated. This can be either a numeric or a factor variable.}
}
\value{
A \code{ggplot} object representing the histogram.
}
\description{
This function generates a histogram for a specified variable in a data frame.
It handles both numeric and factor variables, creating appropriate captions
and titles for the graph based on the characteristics of the variable.
}
\examples{
# Example of using generate_histogram with a numeric variable
dynamic_histogram(ryanhonorthesis, friendship.duration.wks)

# Example of using generate_histogram with a factor variable
dynamic_histogram(ryanhonorthesis, osf_closer_ssf)
}
