% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efa_results.R
\name{efa_apr}
\alias{efa_apr}
\title{Function for Various Tests of Matrix Suitability for Factor Analysis}
\usage{
efa_apr(df)
}
\arguments{
\item{df}{A numeric data frame or matrix where each column represents a variable.
The function will perform the KMO test and Bartlett’s test on this data.}
}
\value{
A list containing:
\describe{
\item{KMO test statistic}{The overall KMO statistic assessing sampling adequacy.}
\item{KMO minimum value}{The minimum MSA (measure of sampling adequacy) value across the variables.}
\item{Bartlett Test of Sphericity P-value}{The p-value from Bartlett's test, used to determine
if the correlation matrix is significantly different from an identity matrix.}
\item{Bartlett Chi-Square Value}{The chi-square value from Bartlett’s test.}
\item{Bartlett Degrees of Freedom}{The degrees of freedom for the Bartlett test.}
\item{KMO Values < .70}{A vector of the variable names where the MSA value is below 0.70, which
indicates that the variables may not be suitable for factor analysis.}
}
}
\description{
This function runs two tests—Kaiser-Meyer-Olkin (KMO) and Bartlett's Test of Sphericity—
to assess the suitability of a correlation matrix for factor analysis (FA).
The KMO test checks the sampling adequacy, and Bartlett's test checks the
hypothesis that the correlation matrix is an identity matrix (which would suggest
that factor analysis is inappropriate).
}
\details{
The KMO test checks the adequacy of the sample for factor analysis. A KMO value greater than 0.70 is considered
acceptable for factor analysis, while a value below 0.50 suggests that the data is not suitable for factor analysis.

Bartlett’s Test of Sphericity tests whether the correlation matrix is an identity matrix, with a significant
result indicating that the variables are inter-correlated and appropriate for factor analysis.

This function is primarily used in exploratory factor analysis (EFA) to assess the suitability of a dataset.
}
\examples{
# Example usage of efa_apr()
efa_apr(ryanhonorthesis[, 2:9])

}
