% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efa_results.R
\name{efa_results}
\alias{efa_results}
\title{Perform and Summarize Exploratory Factor Analysis}
\usage{
efa_results(fa_df, n_factors)
}
\arguments{
\item{fa_df}{A data frame containing the data for factor analysis.}

\item{n_factors}{An integer specifying the number of factors to extract.}
}
\value{
A list containing:
\item{efa}{The EFA model object.}
\item{pm_com}{A combined table of the pattern matrix and communalities formatted for display.}
\item{pm}{The formatted pattern matrix.}
\item{sm}{The formatted structure matrix.}
\item{com_mean}{The mean of communalities.}
\item{com_min}{The minimum communality value.}
\item{com_max}{The maximum communality value.}
\item{coms}{The formatted communalities table.}
}
\description{
This function conducts exploratory factor analysis (EFA) using a maximum likelihood
estimator and summarizes the results, including the pattern matrix, communalities,
and other relevant outputs.
}
\details{
If the number of factors (\code{n_factors}) is 1, additional information about the single-factor
solution is printed. For multi-factor solutions, the function also prints the factor
correlation matrix (if oblique rotation is used).
}
\examples{
# Example usage:
# efa_results(my_data, 3)

}
