% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efa_information.R
\name{fac_alphas}
\alias{fac_alphas}
\title{Calculate Cronbach's Alpha for Items on Each Factor}
\usage{
fac_alphas(df, efa_model, load_cut = 0.4, all_output = FALSE)
}
\arguments{
\item{df}{A data frame containing the observed data, with each column representing an item.}

\item{efa_model}{An EFA model object that contains the factor loadings. This object should have the \code{loadings}
matrix and the number of factors (\code{factors} attribute).}

\item{load_cut}{Numeric value for the factor loading cutoff (default = 0.40). This determines which items are
considered for each alpha calculation}

\item{all_output}{Logical. If \code{TRUE}, the function returns all available output from the \code{psych::alpha()} function,
including detailed item-by-item alpha statistics. If \code{FALSE} (the default), it returns a summarized result with
total alpha and inter-item correlations.}
}
\value{
A list containing results for each factor:
\describe{
\item{\verb{Alpha Totals}}{The overall Cronbach's alpha for the items loading on each factor.}
\item{\verb{Alpha Item Information}}{Detailed item-wise alpha statistics, including drop values.}
\item{\verb{Mean Inter-item Correlation}}{The mean correlation between items on the factor.}
\item{\verb{Minimum/Maximum Inter-item Correlation}}{The minimum and maximum inter-item correlations for the factor.}
}
}
\description{
This function calculates Cronbach's alpha for the items on each factor of an Exploratory Factor Analysis (EFA) model.
The function filters items based on a specified loading cutoff (default loading cutoff = 0.40) and then calculates the
internal consistency of those items using the Cronbach's alpha statistic. The function also calculates and returns
inter-item correlations for each factor.
}
\details{
The \code{fac_alphas} function uses the \code{psych::alpha()} function to calculate Cronbach's alpha, a measure of internal
consistency. The function allows users to filter items based on a factor loading threshold (default = 0.40) and returns
either a full output or a summary, depending on the \code{all_output} argument. The function is particularly useful for
assessing the reliability of scales or factors in factor analysis.
}
\examples{
# Example usage of fac_alphas
# Assuming `efa_model` is a valid EFA model object from a factor analysis
#fac_alphas(df, efa_model)
}
