% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efa_information.R
\name{fac_loadings}
\alias{fac_loadings}
\title{Calculate Number of Items Loading on Each Factor and Factor Loading Statistics}
\usage{
fac_loadings(df, efa_model, load_cut = 0.4)
}
\arguments{
\item{df}{A data frame containing the observed data, with each column representing an item.}

\item{efa_model}{An EFA model object that contains the factor loadings. This object should have the \code{loadings}
matrix and the number of factors (\code{factors} attribute).}

\item{load_cut}{Numeric value for the factor loading cutoff (default = 0.40). This determines which items are
considered as loading on the factor.}
}
\value{
A list containing results for each factor:
\describe{
\item{\verb{Number of Items}}{The number of items loading on the factor above the specified cutoff.}
\item{\code{Items}}{The names of the items that load on the factor above the cutoff.}
\item{\verb{Factor Loading Mean}}{The mean of the factor loadings for the factor.}
\item{\verb{Factor Loading Min/Max}}{The minimum and maximum factor loadings for the factor.}
}
}
\description{
This function calculates the number of items loading on each factor based on a specified cutoff (default = 0.40).
It also computes summary statistics for the factor loadings, including the mean, minimum, and maximum factor loadings
for each factor. The function returns the number of items, the items themselves, and loading statistics for each factor.
}
\details{
The \code{fac_loadings} function filters items based on their factor loadings above a specified cutoff (default = 0.40).
It provides summary statistics for these loadings, including the number of items, as well as the mean, minimum, and
maximum loadings for each factor. This function is useful for understanding which items contribute most to each factor.
}
\examples{
# Example usage of fac_loadings
# Assuming `efa_model` is a valid EFA model object from a factor analysis
#fac_loadings(df, efa_model)
}
