% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_effects_removal.R
\name{feid_set_class}
\alias{feid_set_class}
\title{Tag a history object as \code{fe_int_drop_history}}
\usage{
feid_set_class(x)
}
\arguments{
\item{x}{An object to tag (usually a list produced while dropping fixed-effect interactions).}
}
\value{
\code{x} with class \code{"fe_int_drop_history"} prepended.
}
\description{
Adds the class \code{"fe_int_drop_history"} to an existing object (typically
a list of logged steps) so methods and printers can recognize it.
}
\examples{
H <- list()
H <- feid_set_class(H)
class(H)

}
\keyword{internal}
