% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fivelovelanguages}
\alias{fivelovelanguages}
\title{Five Love Languages Dataset}
\format{
A data frame with 950 rows and 109 variables.
}
\source{
Master's thesis dataset collected in summer/fall 2024 and
cleaned for inclusion in the codechest package.
}
\usage{
fivelovelanguages
}
\description{
A cleaned dataset derived from my master's thesis examining the
psychometric structure of the Five Love Languages.
}
\details{
The dataset includes demographic variables, relationship variables,
and items measuring both the \emph{receiving} and \emph{expressing} dimensions
of the five love languages.
}
\examples{
data(fivelovelanguages)
head(fivelovelanguages)
}
\keyword{datasets}
