% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_and_lmer_helpers.R
\name{grab_dv}
\alias{grab_dv}
\title{Get dependent variable name from a fitted model}
\usage{
grab_dv(model)
}
\arguments{
\item{model}{A fitted model object.}
}
\value{
A length-1 character vector with the dependent variable name.
}
\description{
Extracts the dependent variable (left-hand side) from a model's formula.
}
\keyword{lm_helper}
