% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_themes.R
\name{jermeys_theme}
\alias{jermeys_theme}
\title{Custom ggplot2 Theme: Jeremy's Theme}
\usage{
jermeys_theme(base_size = 24, base_family = "")
}
\arguments{
\item{base_size}{Base font size. Default is 24.}

\item{base_family}{Base font family. Default is an empty string.}
}
\value{
A ggplot2 theme object.
}
\description{
A custom theme inspired by Jeremy Hogeveen for ggplot2 visualizations.
}
\examples{
\dontrun{
library(ggplot2)
ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  jermeys_theme()
}
}
