% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_block_parser.R
\name{parse_level_blocks}
\alias{parse_level_blocks}
\title{Parse a multilevel model string into blocks}
\usage{
parse_level_blocks(model_string)
}
\arguments{
\item{model_string}{A lavaan model-syntax string containing at least one
\verb{level: <name>} line.}
}
\value{
A list of blocks. Each block has \code{lines} (its lines of syntax)
and \code{indices}, which is either \code{list(type = "prelude")} or
\verb{list(type = "level", block_id = <int>, level_name = <chr>)}.
}
\description{
The multilevel equivalent of \code{\link[=parse_group_blocks]{parse_group_blocks()}}: splits \code{model_string}
at each \verb{level: <name>} line into a leading "prelude" block followed by
one block per level. Used as the \code{parser} argument to \code{\link[=build_model]{build_model()}} and
the \verb{set_*()} functions when \code{type = "levels"}.
}
\examples{
parse_level_blocks(level_syntax("f1 =~ x1 + x2 + x3", c(1, 2)))

}
\seealso{
\code{\link[=parse_single_block]{parse_single_block()}}, \code{\link[=parse_group_blocks]{parse_group_blocks()}}

Other lavaan_syntax:
\code{\link[=group_syntax]{group_syntax()}},
\code{\link[=level_syntax]{level_syntax()}},
\code{\link[=parse_group_blocks]{parse_group_blocks()}},
\code{\link[=parse_single_block]{parse_single_block()}}
}
\concept{lavaan_syntax}
