% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_multiverse.R
\name{print.anova_multiverse}
\alias{print.anova_multiverse}
\title{Print a multiverse analysis}
\usage{
\method{print}{anova_multiverse}(x, ...)
}
\arguments{
\item{x}{An \code{anova_multiverse} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a multiverse analysis
}
