% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_describe.R
\name{print.anova_report}
\alias{print.anova_report}
\title{Print an ANOVA report}
\usage{
\method{print}{anova_report}(x, ...)
}
\arguments{
\item{x}{An \code{anova_report} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an ANOVA report
}
