% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ryanhonorthesis}
\alias{ryanhonorthesis}
\title{Ryan Honors Thesis Dataset}
\format{
Data are in wide format where the scenario people are responding to emotions
on is depicted by "osf." ssf." or "rom." as a variable prefix.
\describe{
\item{osf.surprised}{Surprise at opposite-sex friend hanging out with their opposite-sex friend}
\item{ssf.angry}{Anger at opposite-sex friend hanging out with their same-sex friend}
\item{rom.happy}{Happiness at opposite-sex friend hanging out with their romantic parnter}
}
}
\source{
Data collect in fall 2022 for Ryan Dobson's undergrad honor thesis on jealousy in opposite-sex friendships
}
\usage{
ryanhonorthesis
}
\description{
Ryan Honors Thesis Dataset
}
\examples{
data(ryanhonorthesis)
summary(ryanhonorthesis)
}
\keyword{datasets}
