% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_themes.R
\name{theme_black}
\alias{theme_black}
\title{Custom ggplot2 Theme: Black Theme}
\usage{
theme_black(base_size = 24, base_family = "")
}
\arguments{
\item{base_size}{Base font size. Default is 24.}

\item{base_family}{Base font family. Default is an empty string.}
}
\value{
A ggplot2 theme object.
}
\description{
A ggplot2 theme with a black background and white text.
}
\examples{
\dontrun{
library(ggplot2)
ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  theme_black()
}
}
