% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{.norm_vec}
\alias{.norm_vec}
\title{Normalize a vector of fixed-effect term strings}
\usage{
.norm_vec(v)
}
\arguments{
\item{v}{Character vector of terms.}
}
\value{
A character vector of unique normalized terms.
}
\description{
Applies \code{.norm_fe_term()} to each element and returns unique normalized terms.
}
\examples{
\dontrun{
.norm_vec(c("a*b","a * b","a:b"))
 }
}
\seealso{
Other print_helpers: 
\code{\link{.diff_removed}()},
\code{\link{.get_fixed}()},
\code{\link{.model_info}()},
\code{\link{.norm_fe_term}()},
\code{\link{.re_corr_status}()},
\code{\link{.re_terms_from_model}()},
\code{\link{.step_label}()},
\code{\link{.wrapcat}()},
\code{\link{print.fixed_effect_drop_history}()}
}
\concept{print_helpers}
