% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{jermeys_palette}
\alias{jermeys_palette}
\title{Jeremy's Palette}
\format{
A character vector of hexadecimal color codes.
}
\usage{
jermeys_palette
}
\description{
A custom color palette inspired by Jeremy Hogeveen.
}
\examples{
\dontrun{
scales::show_col(jermeys_palette)
}
}
\keyword{datasets}
